# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.Disable2FARequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  dto.Enable2FARequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
...
  dto.Initiate2FAResponse:
    properties:
      qr_code:
        description: base64
        type: string
      recovery_codes:
        items:
          type: string
        type: array
      secret:
        type: string
    type: object
...
  dto.RegenerateCodesRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
...
  dto.Verify2FARequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
host: localhost:8080
info:
  contact:
    email: support@synergyconnect.com
    name: API Support
  description: Университетская социальная сеть с полным набором функций
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://synergyconnect.com/terms
  title: SynergyConnect API
  version: 1.0.0
paths:
  /2fa/disable:
    post:
      consumes:
      - application/json
      description: Отключает 2FA после проверки кода
      parameters:
      - description: Код подтверждения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Disable2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отключение 2FA
      tags:
      - 2fa
  /2fa/enable:
    post:
      consumes:
      - application/json
      description: Подтверждает код и включает 2FA
      parameters:
      - description: Код подтверждения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Enable2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Включение 2FA
      tags:
      - 2fa
  /2fa/initiate:
    post:
      description: Генерирует секрет и QR-код для Google Authenticator
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Initiate2FAResponse'
      security:
      - BearerAuth: []
      summary: Инициализация 2FA
      tags:
      - 2fa
  /2fa/recovery-codes:
    get:
      description: Возвращает коды восстановления
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - BearerAuth: []
      summary: Получение кодов восстановления
      tags:
      - 2fa
  /2fa/recovery-codes/regenerate:
    post:
      consumes:
      - application/json
      description: Генерирует новые коды восстановления
      parameters:
      - description: Код подтверждения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegenerateCodesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - BearerAuth: []
      summary: Перегенерация кодов восстановления
      tags:
      - 2fa
  /2fa/status:
    get:
      description: Проверяет, включена ли 2FA для пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Статус 2FA
      tags:
      - 2fa
  /2fa/verify:
    post:
      consumes:
      - application/json
      description: Проверяет правильность кода без изменения статуса
      parameters:
      - description: Код для проверки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Проверка TOTP кода
      tags:
      - 2fa
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"